@interface/Dbi -> Add sanitize()

+++ utils/interface.ts
@@ -2,6 +2,19 @@
 import { Collection, type CollectionInfo, type DeleteResult, type InsertOneResult, MongoClient, ObjectId, type UpdateResult, type WithId } from 'mongodb'
 import type { ObjectAny } from '../type'
 
+const sanitize = (object: ObjectAny) => {
+    if (object instanceof Object) {
+        for (const key in object) {
+            if (/^\$/.test(key)) {
+                delete object[key];
+            } else {
+                sanitize(object[key]); // for nested
+            }
+        }
+    }
+    return object
+}
+
 export class Dbi {
     db: string
     client: MongoClient
@@ -11,21 +24,32 @@ export class Dbi {
         this.client = client
     }
     async add(collection: string, object: object): Promise<InsertOneResult> {
-        return await this.client.db(this.db).collection(collection).insertOne(object)
+        const safeObj = sanitize(object)
+        return await this.client.db(this.db).collection(collection).insertOne(safeObj)
     }
     async upd(collection: string, objectID: ObjectId, updateObject: object): Promise<UpdateResult> {
+        if (!ObjectId.isValid(objectID)) {
+            throw new Error('Invalid ObjectId');
+        }
+        const safeObj = sanitize(updateObject)
         return await this.client.db(this.db).collection(collection).updateOne({
             _id: objectID
         }, {
-            $set: updateObject
+            $set: safeObj
         })
     }
     async del(collection: string, objectID: ObjectId): Promise<DeleteResult> {
+        if (!ObjectId.isValid(objectID)) {
+            throw new Error('Invalid ObjectId');
+        }
         return await this.client.db(this.db).collection(collection).deleteOne({ _id: objectID })
     }
-    async get(collection: string, queryObject: ObjectAny , quantity: number = 1, sort: boolean = false, skip: number = 0): Promise<WithId<object>[] | null[]> {
+    async get(collection: string, queryObject: ObjectAny, quantity: number = 1, sort: boolean = false, skip: number = 0, doSanitize: boolean = true): Promise<WithId<object>[] | null[]> {
+
+        const safeObj = doSanitize ? sanitize(queryObject) : queryObject
+
         let document: WithId<object>[] | null = [];
-        let cursor = this.client.db(this.db).collection(collection).find(queryObject, {
+        let cursor = this.client.db(this.db).collection(collection).find(safeObj, {
             limit: (quantity === -1) ? undefined : quantity,
             sort: sort ? { _id: -1 } : undefined,
             skip: skip
